remove old named pipe stuff
