+++ IntelPresentMon/PresentMonMiddleware/FrameEventQuery.cpp
@@ -1,7 +1,7 @@
 // Copyright (C) 2017-2024 Intel Corporation
 // SPDX-License-Identifier: MIT
 #define NOMINMAX
-#include "../PresentMonUtils/PresentMonNamedPipe.h"
+#include "../PresentMonUtils/StreamFormat.h"
 #include "FrameEventQuery.h"
 #include "../PresentMonAPIWrapperCommon/Introspection.h"
 #include "../CommonUtilities/Memory.h"
