+++ IntelPresentMon/Streamer/StreamClient.h
@@ -5,7 +5,7 @@
 #include <thread>
 #include <string>
 #include <map>
-#include "../PresentMonUtils/PresentMonNamedPipe.h"
+#include "../PresentMonUtils/StreamFormat.h"
 #include "../PresentMonUtils/LegacyAPIDefines.h"
 #include "NamedSharedMemory.h"
